lavinmq_tutorials/python_tutorials/requirements.txt
pika==1.1.0
python-dotenv 